requirements.txt
onnxruntime>=1.16.0
torch>=2.0.0
torchaudio>=2.0.0
tokenizers>=0.15.0
piper-tts>=1.2.0
faiss-cpu>=1.7.4
sounddevice>=0.4.6
soundfile>=0.12.0
PyPDF2>=3.0.0
python-docx>=1.0.0
numpy>=1.24.0
pyyaml>=6.0
gigaam>=0.1.0
omegaconf>=2.3.0
sentence-transformers>=2.0.0
jiwer>=3.0.0
